requirements.txt
Flask==3.1.2
Werkzeug==3.1.5
Pillow==12.1.1
opencv-python==4.10.0.84
numpy==2.4.2
ultralytics==8.3.157
